[refactor] move UI components to View Controller
