--- ChatBot/ChatBot/Presentation/GPTChatRoom/View/GPTChatRoomView.swift
@@ -1,88 +1,0 @@
-//
-//  GPTChatRoomView.swift
-//  ChatBot
-//
-//  Created by 김준성 on 1/11/24.
-//
-
-import UIKit
-
-final class GPTChatRoomView: UIView {
-    lazy var chatCollectionView: UICollectionView = {
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: configureCollectionViewLayout())
-        collectionView.translatesAutoresizingMaskIntoConstraints = false
-        collectionView.register(GPTChatRoomCell.self, forCellWithReuseIdentifier: "\(type(of: self))")
-        collectionView.backgroundColor = .clear
-        return collectionView
-    }()
-    
-    lazy var horizontalStackView: UIStackView = {
-       let stackView = UIStackView()
-        stackView.axis = .horizontal
-        stackView.translatesAutoresizingMaskIntoConstraints = false
-        stackView.distribution = .fill
-        stackView.addArrangedSubview(commentTextField)
-        stackView.addArrangedSubview(sendButton)
-        return stackView
-    }()
-    
-    lazy var commentTextField: UITextField = {
-        let textField = UITextField()
-        textField.font = UIFont.preferredFont(forTextStyle: .body)
-        textField.borderStyle = .roundedRect
-        textField.backgroundColor = .white
-        textField.translatesAutoresizingMaskIntoConstraints = false
-        textField.setContentCompressionResistancePriority(.defaultLow, for: .horizontal)
-        return textField
-    }()
-    
-    lazy var sendButton: UIButton = {
-        let button = UIButton()
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.setContentHuggingPriority(.defaultHigh, for: .horizontal)
-        button.setContentCompressionResistancePriority(.defaultHigh, for: .horizontal)
-        
-        let imageConfiguration = UIImage.SymbolConfiguration(pointSize: 32)
-        let uiImage = UIImage(systemName: "arrow.up.circle.fill", withConfiguration: imageConfiguration)
-        button.setImage(uiImage, for: .normal)
-        button.contentHorizontalAlignment = .fill
-        button.contentVerticalAlignment = .fill
-        
-        return button
-    }()
-    
-    init() {
-        super.init(frame: .zero)
-        setViews()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func setViews() {
-        addSubview(chatCollectionView)
-        addSubview(horizontalStackView)
-        
-        NSLayoutConstraint.activate([
-            chatCollectionView.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor),
-            chatCollectionView.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor),
-            chatCollectionView.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor),
-            chatCollectionView.bottomAnchor.constraint(equalTo: horizontalStackView.topAnchor),
-            horizontalStackView.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor, constant: 4),
-            horizontalStackView.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor, constant: -4),
-            horizontalStackView.bottomAnchor.constraint(equalTo: keyboardLayoutGuide.topAnchor, constant: -4)
-        ])
-    }
-    
-    private func configureCollectionViewLayout() -> UICollectionViewCompositionalLayout {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(100))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(100))
-        let group = NSCollectionLayoutGroup.vertical(layoutSize: groupSize, subitems: [item])
-        
-        let section = NSCollectionLayoutSection(group: group)
-        return UICollectionViewCompositionalLayout(section: section)
-    }
-}
