+++ ChatBot/ChatBot/Presentation/GPTChatRoom/ViewController/GPTChatRoomViewController.swift
@@ -13,14 +13,55 @@ final class GPTChatRoomViewController: UIViewController {
         case main
     }
     
+    private lazy var chatCollectionView: UICollectionView = {
+        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: configureCollectionViewLayout())
+        collectionView.translatesAutoresizingMaskIntoConstraints = false
+        collectionView.register(GPTChatRoomCell.self, forCellWithReuseIdentifier: "\(type(of: GPTChatRoomCell.self))")
+        return collectionView
+    }()
+    
+    private lazy var horizontalStackView: UIStackView = {
+        let stackView = UIStackView()
+         stackView.axis = .horizontal
+         stackView.translatesAutoresizingMaskIntoConstraints = false
+         stackView.distribution = .fill
+         stackView.addArrangedSubview(commentTextField)
+         stackView.addArrangedSubview(sendButton)
+         return stackView
+     }()
+    
+    private lazy var commentTextField: UITextField = {
+        let textField = UITextField()
+        textField.font = UIFont.preferredFont(forTextStyle: .body)
+        textField.borderStyle = .roundedRect
+        textField.backgroundColor = .white
+        textField.translatesAutoresizingMaskIntoConstraints = false
+        textField.setContentCompressionResistancePriority(.defaultLow, for: .horizontal)
+        return textField
+    }()
+    
+    private lazy var sendButton: UIButton = {
+        let button = UIButton()
+        button.translatesAutoresizingMaskIntoConstraints = false
+        button.setContentHuggingPriority(.defaultHigh, for: .horizontal)
+        button.setContentCompressionResistancePriority(.defaultHigh, for: .horizontal)
+        
+        let imageConfiguration = UIImage.SymbolConfiguration(pointSize: 32)
+        let uiImage = UIImage(systemName: "arrow.up.circle.fill", withConfiguration: imageConfiguration)
+        button.setImage(uiImage, for: .normal)
+        button.contentHorizontalAlignment = .fill
+        button.contentVerticalAlignment = .fill
+        button.addTarget(nil, action: #selector(tapSendButton(_:)), for: .touchUpInside)
+        return button
+    }()
+    
     private let viewModel: GPTChatRoomVMProtocol
     private let cellResistration = UICollectionView.CellRegistration<GPTChatRoomCell, Model.GPTMessage> { cell, indexPath, itemIdentifier in
         cell.configureCell(to: itemIdentifier)
     }
     
-    private var chatRoomView: GPTChatRoomView!
-    private var cancellables = Set<AnyCancellable>()
     private var chattingDataSource: UICollectionViewDiffableDataSource<Section, Model.GPTMessage>!
+    private var cancellables = Set<AnyCancellable>()
     
     init(viewModel: GPTChatRoomVMProtocol) {
         self.viewModel = viewModel
@@ -33,10 +74,8 @@ final class GPTChatRoomViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        chatRoomView = GPTChatRoomView()
-        chatRoomView.sendButton.addTarget(nil, action: #selector(tapSendButton(_:)), for: .touchUpInside)
-        view = chatRoomView
-        
+        view.backgroundColor = .systemBackground
+        setViews()
         bind()
         configureDataSource()
     }
@@ -54,26 +93,64 @@ final class GPTChatRoomViewController: UIViewController {
                 
                 if !messages.isEmpty {
                     let indexPath = IndexPath(item: messages.count - 1, section: 0)
-                    self.chatRoomView.chatCollectionView.scrollToItem(at: indexPath, at: .top, animated: true)
+                    chatCollectionView.scrollToItem(at: indexPath, at: .top, animated: true)
                 }
             }
             .store(in: &cancellables)
     }
-    
+}
+
+// MARK: - set Views
+extension GPTChatRoomViewController {
+    private func setViews() {
+        view.addSubview(chatCollectionView)
+        view.addSubview(horizontalStackView)
+        
+        NSLayoutConstraint.activate([
+            chatCollectionView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
+            chatCollectionView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
+            chatCollectionView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
+            chatCollectionView.bottomAnchor.constraint(equalTo: horizontalStackView.topAnchor),
+            horizontalStackView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 4),
+            horizontalStackView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -4),
+            horizontalStackView.bottomAnchor.constraint(equalTo: view.keyboardLayoutGuide.topAnchor, constant: -4)
+        ])
+    }
+}
+
+// MARK: - configure Collection View
+extension GPTChatRoomViewController {
+    private func configureCollectionViewLayout() -> UICollectionViewCompositionalLayout {
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(100))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(100))
+        let group = NSCollectionLayoutGroup.vertical(layoutSize: groupSize, subitems: [item])
+        
+        let section = NSCollectionLayoutSection(group: group)
+        return UICollectionViewCompositionalLayout(section: section)
+    }
+}
+
+// MARK: - configure Diffable Data Source
+extension GPTChatRoomViewController {
     private func configureDataSource() {
         chattingDataSource = UICollectionViewDiffableDataSource<Section, Model.GPTMessage>(
-            collectionView: chatRoomView.chatCollectionView,
+            collectionView: chatCollectionView,
             cellProvider: { collectionView, indexPath, itemIdentifier in
                 collectionView.dequeueConfiguredReusableCell(using: self.cellResistration, for: indexPath, item: itemIdentifier)
             }
         )
     }
-    
+}
+
+ // MARK: - set UIRespond
+extension GPTChatRoomViewController {
     @objc
     private func tapSendButton(_ sender: Any) {
-        if let content = chatRoomView.commentTextField.text, !content.isEmpty {
+        if let content = commentTextField.text, !content.isEmpty {
             viewModel.sendComment(Model.UserMessage(content: content))
         }
-        chatRoomView.commentTextField.text = nil
+        commentTextField.text = nil
     }
 }
